refactor(ioxide): split the HTTP/1.1 loop out of ConnectionDriver, drop the duplicated StatusLine
ConnectionDriver had grown into two unrelated jobs: deciding what protocol a TCP
connection speaks, and then serving it when the answer was HTTP/1.1. The second
half moves to Http1Driver, alongside Http2Driver and Http3Driver - so each
protocol is one file, and ConnectionDriver is only the transport plus the
ALPN/preface decision that routes to them.

The engine also carried its own copy of StatusLine, byte-identical to the one in
GenHTTP.Engine.Shared.Types. Use the shared one; the Ioxide engine gets the same
InternalsVisibleTo the acceptance tests already have.

DateHeader stays duplicated on purpose - the engine's is [ThreadStatic] so each
reactor owns its buffer, which the shared static cannot be.
