+++ Engine/Ioxide/GenHTTP.Engine.Ioxide.csproj
@@ -17,7 +17,7 @@
     <PackageReference Include="ioxide.ngtcp2" Version="0.4.186" />
     <PackageReference Include="ioxide.nghttp3" Version="0.4.186" />
     <PackageReference Include="Glyph11" Version="0.3.6" />
-    <!-- Parallel parse path for benchmarking — GENHTTP_IOXIDE_PARSER=pico (see ConnectionDriver). -->
+    <!-- Parallel parse path for benchmarking — GENHTTP_IOXIDE_PARSER=pico (see Http1Driver). -->
     <PackageReference Include="Glyph11.Pico" Version="0.0.1" />
   </ItemGroup>
