+++ Engine/Shared/GenHTTP.Engine.Shared.csproj
@@ -26,6 +26,7 @@
 
     <ItemGroup>
         <InternalsVisibleTo Include="GenHTTP.Testing.Acceptance" />
+        <InternalsVisibleTo Include="GenHTTP.Engine.Ioxide" />
     </ItemGroup>
 
 </Project>
